+++ apps/switchboard/src/index.ts
@@ -1,4 +1,4 @@
-import { ReactorRouterManager } from "@powerhousedao/reactor-api";
+import { Db, SubgraphManager, getDbClient } from "@powerhousedao/reactor-api";
 import { DocumentDriveServer } from "document-drive";
 import * as DocumentModelsLibs from "document-model-libs/document-models";
 import { DocumentModel } from "document-model/document";
@@ -7,8 +7,7 @@ import dotenv from "dotenv";
 import express from "express";
 import http from "http";
 import path from "path";
-import { Pool } from "pg";
-import * as authListener from "./subgraphs/auth";
+import { Knex } from "knex";
 dotenv.config();
 
 // start document drive server with all available document models
@@ -27,14 +26,15 @@ const main = async () => {
     // init drive server
     await driveServer.initialize();
 
-    const pool = new Pool({
-      connectionString: process.env.DATABASE_URL,
-    });
-    await pool.connect();
-    const reactorRouterManager = new ReactorRouterManager(
+    const dbPath = process.env.DATABASE_URL ?? undefined;
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-assignment
+    const knex = getDbClient(dbPath);
+    const reactorRouterManager = new SubgraphManager(
       "/",
       app,
       driveServer,
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
+      knex,
     );
     // init router
     await reactorRouterManager.init();