+++ packages/reactor-api/src/processors/analytics-processor.ts
@@ -1,7 +1,7 @@
 import { IAnalyticsStore } from "@powerhousedao/analytics-engine-core";
 import { Document, OperationScope } from "document-model/document";
 import { Processor } from "./processor";
-import { ProcessorOptions, ProcessorSetupArgs } from "src/types";
+import { ProcessorOptions, ProcessorSetupArgs } from "../types";
 
 export * from "@powerhousedao/analytics-engine-core";
 
@@ -13,11 +13,6 @@ export abstract class AnalyticsProcessor<
 
   constructor(args: ProcessorSetupArgs, options?: ProcessorOptions) {
     super(args, options);
-    this.analyticsStore = args.dataSources.analyticsStore;
-  }
-
-  onSetup(args: ProcessorSetupArgs) {
-    super.onSetup(args);
-    this.analyticsStore = args.dataSources.analyticsStore;
+    this.analyticsStore = args.analyticsStore;
   }
 }