+++ packages/reactor-api/src/subgraphs/manager.ts
@@ -1,45 +1,48 @@
 import { ApolloServer } from "@apollo/server";
 import { expressMiddleware } from "@apollo/server/express4";
 import { ApolloServerPluginInlineTraceDisabled } from "@apollo/server/plugin/disabled";
-import { GraphQLResolverMap } from "@apollo/subgraph/dist/schema-helper";
+import { IAnalyticsStore } from "@powerhousedao/analytics-engine-core";
 import bodyParser from "body-parser";
 import cors from "cors";
 import { IDocumentDriveServer } from "document-drive";
 import express, { IRouter, Router } from "express";
-import { analyticsSubgraph, driveSubgraph, systemSubgraph } from "./subgraphs";
-import { Context, Subgraph } from "./types";
-import { createSchema } from "./utils/create-schema";
+import { Db } from "src/types";
+import { Context, SubgraphArgs, SubgraphClass } from ".";
+import { createSchema } from "../utils/create-schema";
+import { AnalyticsSubgraph } from "./analytics";
+import { Subgraph } from "./base";
+import { DriveSubgraph } from "./drive";
+import { SystemSubgraph } from "./system";
 
-export class ReactorRouterManager {
+export class SubgraphManager {
   private reactorRouter: IRouter = Router();
   private contextFields: Record<string, any> = {};
-
-  // @todo: need to persist somewhere
-  private subgraphs: Subgraph[] = [
-    {
-      name: "system",
-      resolvers: systemSubgraph.resolvers,
-      typeDefs: systemSubgraph.typeDefs,
-    },
-    {
-      name: "d/:drive",
-      resolvers: driveSubgraph.resolvers,
-      typeDefs: driveSubgraph.typeDefs,
-    },
-    {
-      name: "analytics",
-      resolvers: analyticsSubgraph.resolvers as GraphQLResolverMap<Context>,
-      typeDefs: analyticsSubgraph.typeDefs,
-    },
-  ];
+  private subgraphs: Subgraph[] = [];
 
   constructor(
     private readonly path: string,
     private readonly app: express.Express,
     private readonly reactor: IDocumentDriveServer,
-  ) {}
+    private readonly operationalStore: Db,
+    private readonly analyticsStore: IAnalyticsStore,
+  ) {
+    const args: SubgraphArgs = {
+      reactor: this.reactor,
+      operationalStore: this.operationalStore,
+      analyticsStore: this.analyticsStore,
+      subgraphManager: this,
+    };
+
+    // Setup Default subgraphs
+    this.registerSubgraph(SystemSubgraph);
+    this.registerSubgraph(DriveSubgraph);
+    this.registerSubgraph(AnalyticsSubgraph);
+  }
 
   async init() {
+    console.log(
+      `Initializing ReactorRouterManager with subgraphs: [${this.subgraphs.map((e) => e.name).join(", ")}]`,
+    );
     const models = this.reactor.getDocumentModels();
     const driveModel = models.find(
       (it) => it.documentModel.name === "DocumentDrive",
@@ -67,11 +70,10 @@ export class ReactorRouterManager {
     for (const subgraph of this.subgraphs) {
       const subgraphConfig = this.#getLocalSubgraphConfig(subgraph.name);
       if (!subgraphConfig) continue;
-      console.log(`Setting up subgraph ${subgraphConfig.name}`);
       // get schema
       const schema = createSchema(
         this.reactor,
-        subgraphConfig.resolvers as GraphQLResolverMap<Context>,
+        subgraphConfig.resolvers,
         subgraphConfig.typeDefs,
       );
       // create apollo server
@@ -94,7 +96,7 @@ export class ReactorRouterManager {
             headers: req.headers,
             driveId: req.params.drive ?? undefined,
             driveServer: this.reactor,
-            // analyticStore: undefined, // TODO: add analytic store
+            db: this.operationalStore,
             ...this.getAdditionalContextFields(),
           }),
         }),
@@ -102,13 +104,19 @@ export class ReactorRouterManager {
     }
 
     this.reactorRouter = newRouter;
-    console.log("Router updated.");
   }
 
-  async registerSubgraph(subgraph: Subgraph) {
-    this.subgraphs.unshift(subgraph);
-    console.log(`Registered [${subgraph.name}] subgraph.`);
-    return Promise.resolve();
+  async registerSubgraph(subgraph: SubgraphClass) {
+    const subgraphInstance = new subgraph({
+      operationalStore: this.operationalStore,
+      analyticsStore: this.analyticsStore,
+      reactor: this.reactor,
+      subgraphManager: this,
+    });
+    await subgraphInstance.onSetup();
+    this.subgraphs.unshift(subgraphInstance);
+    console.log(`> Registered ${subgraphInstance.name} subgraph.`);
+    await this.updateRouter();
   }
 
   #getLocalSubgraphConfig(subgraphName: string) {